added support for synchronized devices by introducing the MultiDeviceRole (standalone, leader, follower) that can be configured in the rgbd_control. Only kinect azur driver supports leader and follower role. As with the kinect azur viewer one can start rgbd_control one per attached device and configure leader role for the device that is used as master and only has an audio cable plugged into the Synch Out plug and the follower role for all other devices. All followers need to be started first and actually stream data as soons as the leader has been started.

+++ libs/rgbd_capture/rgbd_device_emulation.cxx
@@ -304,7 +304,7 @@ namespace rgbd {
 		}
 
 		double inv_fps = 1000.f / stream->fps;
-		double current_frame_time = chrono::duration_cast<milliseconds>(chrono::steady_clock::now().time_since_epoch()).count();
+		double current_frame_time = (double)chrono::duration_cast<milliseconds>(chrono::steady_clock::now().time_since_epoch()).count();
 		
 		//limit fps
 		if (current_frame_time < *last_frame_time+inv_fps) {
@@ -340,7 +340,7 @@ namespace rgbd {
 			return false;
 		}
 
-		frame.buffer_size = data.size();
+		frame.buffer_size = (unsigned)data.size();
 		if (frame.frame_data.size() != frame.buffer_size) {
 			frame.frame_data.resize(frame.buffer_size);
 		}
